:construction: set up definition for Recipe
Recipe is an object. If you have an object, you need to list the properties such as id, idPublic, name, chef, stepList, ingredients. The properties need to declare their type, example: id: type: number. Sometimes the properties have a type array. If so, you need to list the items in that array and declare the type of the items. If your property has an object, you need to list the properties inside of that object and declare its type.