:construction: create endpoint called /recipes & :sparkles: GET request
I made a route called /recipes. This route grabs the recipes file in my controllers folder in vscode. This is a get request. Description describes the get request. operationId is the function created in the controller. I create 2 responses 200 and 500. 200 has a description and schema. The schema is an array and the items in the schema  pull in the Recipe definition. 500 has a description and a schema. the schema is an object with a property message who's value is a string.