:sparkles: create function getRecipe()
Tested in postman. getRecipe is a function that takes in a req and res. We created an empty array called responseRecipeList that is empty. we created a for loop that checks for each recipe in the array of the function recipeList to see if its isPublic. if it's public, then return the response and turn the responseRecipeList into json and send a stauts of 200