:tada::recycle: refactor code to incorporate MVC design (GET request)
This is to follow MVC designs. There is no views yet. This update if for the GET request. Now we have a controller and model for Recipe. In the recipes controller, getRecipe() has a variable called recipes which was made from a new instance of Recipe (created in the the models folder). This new instance of Recipe aka recipes gets a list of recipes by calling its own function getRecipeList(). recipeList is a function in Recipe Model that just returns hard code of a list of recipes. Back in the controller, we want a list of recipes that has isPublic true. to get that, create an empty array called publicRecipeList. create a for loop that goes over each recipe in recipes. if recipe isPublic (true), then push the true recipe onto the publicRecipeList array. Then take that array and turn it into json and send a status of 200